+++ src/utils/get-timezone-offset-in-hours.js
@@ -1,0 +1,7 @@
+const getTimezoneOffsetInHours = () => {
+  const offsetInMinutes = new Date().getTimezoneOffset()
+  const offsetInHours = -offsetInMinutes / 60
+  return offsetInHours
+}
+
+export { getTimezoneOffsetInHours }
